refactor(okx): extract balance-change check into a generic method
Move the duplicate-balance-update detection out of okx's handleBalance into
BaseExchange.isBalanceChanged so other exchanges can reuse it, and move the
OKX-specific watchBalance assertions out of the shared test into a dedicated
exchange-specific test file.
